Editando tela de produto

+++ OdinPizzaria/src/model/Produto.java
@@ -2,5 +2,5 @@ package model;
 
 public class Produto {
 	public String nome;
-	public float valor;
+	public double valor;
 }
